feat: 재료 조회 및 생성, 재료 카테고리 조회 기능 추가

+++ src/main/java/com/recipe/app/src/ingredient/infra/IngredientCategoryJpaRepository.java
@@ -2,6 +2,6 @@ package com.recipe.app.src.ingredient.infra;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IngredientCategoryJpaRepository extends JpaRepository<IngredientCategoryEntity, Integer> {
+public interface IngredientCategoryJpaRepository extends JpaRepository<IngredientCategoryEntity, Long> {
 
 }
